minor refactors
