+++ HotelAccounting/Model/Employee.cs
@@ -1,0 +1,13 @@
+namespace HotelAccounting
+{
+    public class Employee
+    {
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Phone { get; set; }
+
+        public string Sex { get; set; }
+    }
+}
