+++ HotelAccounting/Model/Guest.cs
@@ -1,0 +1,15 @@
+﻿namespace HotelAccounting
+{
+    public class Guest
+    {
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Phone { get; set; }
+
+        public string Sex { get; set; }
+
+        public string? Photo { get; set; }
+    }
+}
